moved get & generate random int funcs to separate files
