+++ js/utils/generate-random-int.js
@@ -1,0 +1,13 @@
+/**
+ * Get a random integer in range inclusively from 'min' and 'max'.
+ *
+ * @param {number} min - min number
+ * @param {number} max - max number
+ * @return {number} a random integer
+ */
+
+const generateRandomInt = (min, max) => (
+  Math.floor(Math.random() * (max - min + 1) + min)
+);
+
+export default generateRandomInt;
